+++ WebApplication/src/orthancApi.js
@@ -668,6 +668,9 @@ export default {
     async deleteWorklist(worklistId) {
         axios.delete(orthancApiUrl + "worklists/" + worklistId);
     },
+    async createWorklist(wlTags) {
+        return (await axios.post(orthancApiUrl + "worklists/create", {"Tags": wlTags})).data;
+    },
 
     ////////////////////////////////////////// HELPERS
     getOsimisViewerUrl(level, resourceOrthancId) {